frontend_/Scripts/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn==0.23.2

# LangChain et intégrations
langchain==0.1.13
langchain-core==0.1.32
langchain-community==0.0.28
langchain-mistralai==0.1.5  

# Outils de parsing PDF
PyPDF2==3.0.1
python-docx==1.1.0  

# Display dans notebooks
ipython==8.21.0
IPython==8.21.0
notebook==7.1.2
jupyterlab==4.1.5

# .env file
python-dotenv==1.0.1

markdown==3.6
